Put the previous command on players rather than the root object
Allows for an arbitrary number of players in the game.
